requirements.txt
django
djangorestframework
django-cors-headers
dj-rest-auth
django-allauth
requests